+++ blocks/init/src/Blocks/components/image/components/image-editor.js
@@ -1,4 +1,10 @@
-import { checkAttr, getBreakpointData, getBreakpointNames, getTwClasses, getTwPart } from '@eightshift/frontend-libs-tailwind/scripts';
+import {
+	checkAttr,
+	getBreakpointData,
+	getBreakpointNames,
+	getTwClasses,
+	getTwPart,
+} from '@eightshift/frontend-libs-tailwind/scripts';
 import { ImagePlaceholder } from '@eightshift/ui-components';
 import manifest from './../manifest.json';