+++ scripts/components/settings/settings.js
@@ -10,7 +10,11 @@ import { useThemeOptions } from './use-theme-options';
 
 export const EsThemeOptionsContext = createContext(null);
 
-export const ThemeOptionsPage = ({ title = __('Theme options', 'eightshift-frontend-libs-tailwind'), children, settingName = 'eightshift-theme-options' }) => {
+export const ThemeOptionsPage = ({
+	title = __('Theme options', 'eightshift-frontend-libs-tailwind'),
+	children,
+	settingName = 'eightshift-theme-options',
+}) => {
 	const themeOptions = useThemeOptions(settingName);
 	const { isLoading, saveSettings } = themeOptions;
 
@@ -26,7 +30,9 @@ export const ThemeOptionsPage = ({ title = __('Theme options', 'eightshift-front
 						disabled={isLoading}
 						icon={isLoading ? icons.moreH : icons.save}
 					>
-						{isLoading ? __('Saving...', 'eightshift-frontend-libs-tailwind') : __('Save', 'eightshift-frontend-libs-tailwind')}
+						{isLoading
+							? __('Saving...', 'eightshift-frontend-libs-tailwind')
+							: __('Save', 'eightshift-frontend-libs-tailwind')}
 					</Button>
 				</HStack>