+++ scripts/editor/registration.js
@@ -84,7 +84,11 @@ export const registerBlocks = (
 
 			// Get Block Transforms component from block name and transformsComponentPath.
 			if (transformsComponentPath !== null) {
-				const blockTransformsComponent = getBlockGenericComponent(blockManifest.blockName, transformsComponentPath, 'transforms');
+				const blockTransformsComponent = getBlockGenericComponent(
+					blockManifest.blockName,
+					transformsComponentPath,
+					'transforms',
+				);
 
 				if (blockTransformsComponent !== null) {
 					blockManifest.transforms = blockTransformsComponent;
@@ -93,7 +97,11 @@ export const registerBlocks = (
 
 			// Get Block Deprecations component from block name and deprecationsComponentPath.
 			if (deprecationsComponentPath !== null) {
-				const blockDeprecationsComponent = getBlockGenericComponent(blockManifest.blockName, deprecationsComponentPath, 'deprecations');
+				const blockDeprecationsComponent = getBlockGenericComponent(
+					blockManifest.blockName,
+					deprecationsComponentPath,
+					'deprecations',
+				);
 
 				if (blockDeprecationsComponent !== null) {
 					blockManifest.deprecated = blockDeprecationsComponent;
@@ -111,7 +119,11 @@ export const registerBlocks = (
 
 			// Get Block Overrides component from block name and overridesComponentPath.
 			if (overridesComponentPath !== null) {
-				const blockOverridesComponent = getBlockGenericComponent(blockManifest.blockName, overridesComponentPath, 'overrides');
+				const blockOverridesComponent = getBlockGenericComponent(
+					blockManifest.blockName,
+					overridesComponentPath,
+					'overrides',
+				);
 
 				if (blockOverridesComponent !== null) {
 					blockManifest = Object.assign(blockManifest, blockOverridesComponent);
@@ -119,7 +131,14 @@ export const registerBlocks = (
 			}
 
 			// Pass data to registerBlock helper to get final output for registerBlockType.
-			const blockDetails = registerBlock(globalManifest, wrapperManifest, componentsManifest, blockManifest, wrapperComponent, blockComponent);
+			const blockDetails = registerBlock(
+				globalManifest,
+				wrapperManifest,
+				componentsManifest,
+				blockManifest,
+				wrapperComponent,
+				blockComponent,
+			);
 
 			// Format the 'deprecated' attribute details to match the format Gutenberg wants.
 			if (blockDetails?.options?.deprecated) {
@@ -145,7 +164,10 @@ export const registerBlocks = (
 							};
 						},
 
-						isEligible: deprecation?.isEligible ?? ((attributes) => Object.keys(deprecation.oldAttributes).every((v) => Object.keys(attributes).includes(v))),
+						isEligible:
+							deprecation?.isEligible ??
+							((attributes) =>
+								Object.keys(deprecation.oldAttributes).every((v) => Object.keys(attributes).includes(v))),
 						save: blockDetails.options.save,
 					};
 				});
@@ -196,7 +218,12 @@ export const registerBlocks = (
  * );
  * ```
  */
-export const registerVariations = (globalManifest = {}, variationsManifestPath, blocksManifestPath = null, overridesComponentPath = null) => {
+export const registerVariations = (
+	globalManifest = {},
+	variationsManifestPath,
+	blocksManifestPath = null,
+	overridesComponentPath = null,
+) => {
 	const variationsManifests = variationsManifestPath.keys().map(variationsManifestPath);
 
 	// Set all store values.
@@ -210,7 +237,11 @@ export const registerVariations = (globalManifest = {}, variationsManifestPath, 
 		if (active) {
 			// Get Block Overrides component from block name and overridesComponentPath.
 			if (overridesComponentPath !== null) {
-				const blockOverridesComponent = getBlockGenericComponent(variationManifest.name, overridesComponentPath, 'overrides');
+				const blockOverridesComponent = getBlockGenericComponent(
+					variationManifest.name,
+					overridesComponentPath,
+					'overrides',
+				);
 
 				if (blockOverridesComponent !== null) {
 					variationManifest = Object.assign(variationManifest, blockOverridesComponent);
@@ -218,7 +249,11 @@ export const registerVariations = (globalManifest = {}, variationsManifestPath, 
 			}
 
 			// Pass data to registerVariation helper to get final output for registerBlockVariation.
-			const blockDetails = registerVariation(globalManifest, variationManifest, blocksManifestPath !== null ? blocksManifestPath.keys().map(blocksManifestPath) : []);
+			const blockDetails = registerVariation(
+				globalManifest,
+				variationManifest,
+				blocksManifestPath !== null ? blocksManifestPath.keys().map(blocksManifestPath) : [],
+			);
 
 			// Native WP method for block registration.
 			registerBlockVariation(blockDetails.blockName, blockDetails.options);
@@ -248,11 +283,15 @@ export const getBlockEditComponent = (blockName, paths, fileName) => {
 	const pathsKeys = paths.keys();
 
 	// Get Block edit component from block name and pathsKeys.
-	const editComponent = pathsKeys.filter((filePath) => filePath === `./${blockName}/${blockName}-${fileName}.js`).map(paths)[0];
+	const editComponent = pathsKeys
+		.filter((filePath) => filePath === `./${blockName}/${blockName}-${fileName}.js`)
+		.map(paths)[0];
 
 	// If edit component is missing throw and error.
 	if (typeof editComponent === 'undefined') {
-		throw Error(`It looks like you are missing block edit component for block: ${blockName}, please check if you have ${blockName}-block.js file in your block folder.`);
+		throw Error(
+			`It looks like you are missing block edit component for block: ${blockName}, please check if you have ${blockName}-block.js file in your block folder.`,
+		);
 	}
 
 	// No mater if class of functional component is used fetch the first item in an object.
@@ -260,7 +299,9 @@ export const getBlockEditComponent = (blockName, paths, fileName) => {
 
 	// If edit component callback is missing throw and error.
 	if (typeof editCallback === 'undefined') {
-		throw Error(`It looks like you are missing block edit component for block: ${blockName}, please check if you have ${blockName}-block.js file in your block folder.`);
+		throw Error(
+			`It looks like you are missing block edit component for block: ${blockName}, please check if you have ${blockName}-block.js file in your block folder.`,
+		);
 	}
 
 	return editCallback;
@@ -283,7 +324,9 @@ export const getBlockGenericComponent = (blockName, paths, fileName) => {
 	const pathsKeys = paths.keys();
 
 	// Get Block edit component from block name and pathsKeys.
-	const editComponent = pathsKeys.filter((filePath) => filePath === `./${blockName}/${blockName}-${fileName}.js`).map(paths)[0];
+	const editComponent = pathsKeys
+		.filter((filePath) => filePath === `./${blockName}/${blockName}-${fileName}.js`)
+		.map(paths)[0];
 
 	// If edit component is missing throw and error.
 	if (typeof editComponent === 'undefined') {
@@ -492,7 +535,14 @@ export const getIconOptions = (globalManifest, blockManifest) => {
  *
  * @returns {object}
  */
-export const prepareComponentAttribute = (manifest, newName, realName, isExample = false, parent = '', currentAttributes = false) => {
+export const prepareComponentAttribute = (
+	manifest,
+	newName,
+	realName,
+	isExample = false,
+	parent = '',
+	currentAttributes = false,
+) => {
 	const output = {};
 
 	// Define different data point for attributes or example.
@@ -563,17 +613,30 @@ export const prepareComponentAttributes = (componentsManifest, manifest, isExamp
 
 		// Bailout if component doesn't exist.
 		if (!component) {
-			throw Error(`Component specified in "${name}" manifest doesn't exist in your components list. Please check if you project has "${realComponentName}" component.`);
+			throw Error(
+				`Component specified in "${name}" manifest doesn't exist in your components list. Please check if you project has "${realComponentName}" component.`,
+			);
 		}
 
 		let outputAttributes = {};
 
 		// If component has more components do recursive loop.
 		if (component?.components) {
-			outputAttributes = prepareComponentAttributes(componentsManifest, component, isExample, `${newParent}${upperFirst(camelCase(newComponentName))}`);
+			outputAttributes = prepareComponentAttributes(
+				componentsManifest,
+				component,
+				isExample,
+				`${newParent}${upperFirst(camelCase(newComponentName))}`,
+			);
 		} else {
 			// Output the component attributes if there is no nesting left, and append the parent prefixes.
-			outputAttributes = prepareComponentAttribute(component, newComponentName, realComponentName, isExample, newParent);
+			outputAttributes = prepareComponentAttribute(
+				component,
+				newComponentName,
+				realComponentName,
+				isExample,
+				newParent,
+			);
 		}
 
 		// Populate the output recursively.
@@ -752,7 +815,14 @@ export const registerVariation = (globalManifest = {}, variationManifest = {}) =
  *
  * @returns {object}
  */
-export const registerBlock = (globalManifest = {}, wrapperManifest = {}, componentsManifest = {}, blockManifest = {}, wrapperComponent, blockComponent) => {
+export const registerBlock = (
+	globalManifest = {},
+	wrapperManifest = {},
+	componentsManifest = {},
+	blockManifest = {},
+	wrapperComponent,
+	blockComponent,
+) => {
 	// Block Icon option.
 	blockManifest['icon'] = getIconOptions(globalManifest, blockManifest);