+++ scripts/helpers/index.js
@@ -1,5 +1,11 @@
 // All exports are sorted in alphabetical order.
 
-export { getBreakpointData, getBreakpointNames, getGlobalManifest, getBreakpointUiData, getResponsiveData } from './breakpoints';
+export {
+	getBreakpointData,
+	getBreakpointNames,
+	getGlobalManifest,
+	getBreakpointUiData,
+	getResponsiveData,
+} from './breakpoints';
 export { cookies } from './cookies';
 export { dynamicImport } from './dynamic-import';